requirements.txt
hikka-tl-new==2.0.4
hikka-pyro-new==2.0.103
aiofile==3.8.1
gitpython==3.1.27
meval==2.5
Jinja2==3.0.3
requests==2.27.1
aiogram>=3.3.0
grapheme==0.6.0
emoji==2.1.0
werkzeug==2.2.2
psutil==5.9.4
tgcrypto==1.2.5
rsa==4.9
ruamel.yaml==0.17.21


uvloop==0.19.0
orjson==3.9.15
aiosqlite==0.19.0
aiofiles==23.2.1
fastapi
uvicorn
python-multipart
